moving the black style to its own file
